pyproject.toml
[tool.poetry]
name = "servergrimoiredjango"
version = "0.1.1"
description = "A sysadmin module for django"
authors = ["Fundor333 <fundor333@fundor333.com>"]
license = "Mit"
keywords = ["devops", "ssl", "sysadmin", "django"]

classifiers = [
    'Intended Audience :: System Administrators',
    'License :: OSI Approved :: MIT License',
    'Topic :: System :: Networking :: Monitoring',
    'Intended Audience :: Developers',
    "Development Status :: 4 - Beta",
    "Framework :: Django",
]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.0"
django-colorfield = "^0.3.2"
coverage = "^5.3"
pyOpenssl = "^19.1.0"
dnspython = "^2.0.0"
python-whois = "^0.7.3"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
pylint-django = "^2.3.0"
pre-commit = "^2.7.1"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
pytest-django = "^4.1.0"
skjold = "^0.1.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.skjold]
sources = ["pyup", "gemnasium"]  # Sources to check against.
report_only = true                         # Report only, always exit with zero.
cache_dir = '.skylt_cache'                 # Cache location (default: `~/.skjold/cache`).
cache_expires = 86400                      # Cache max. age.
verbose = true                             # Be verbose.
